Add tests for saml registered services in spa service registry

+++ core/cas-server-core-tickets/src/main/java/org/apereo/cas/ticket/support/NeverExpiresExpirationPolicy.java
@@ -35,13 +35,13 @@ public class NeverExpiresExpirationPolicy extends AbstractCasExpirationPolicy {
     @JsonIgnore
     @Override
     public Long getTimeToLive() {
-        return Long.MAX_VALUE;
+        return (long) Integer.MAX_VALUE;
     }
 
     @JsonIgnore
     @Override
     public Long getTimeToIdle() {
-        return Long.MAX_VALUE;
+        return (long) Integer.MAX_VALUE;
     }
     
     @Override
